+++ generator/lib/src/field_element_ex.dart
@@ -17,9 +17,11 @@
 ////////////////////////////////////////////////////////////////////////////////
 import 'package:analyzer/dart/ast/ast.dart';
 import 'package:analyzer/dart/element/element.dart';
+import 'package:analyzer/dart/element/type.dart';
 import 'package:build/build.dart';
 import 'package:realm_common/realm_common.dart';
 import 'package:realm_generator/src/expanded_context_span.dart';
+import 'package:realm_generator/src/pseudo_type.dart';
 import 'package:source_gen/source_gen.dart';
 import 'package:source_span/source_span.dart';
 
@@ -55,17 +57,11 @@ extension FieldElementEx on FieldElement {
         [span!],
       );
 
-  // Works even if type of field is unresolved
-  String get typeText => (typeAnnotation ?? initializerExpression?.staticType ?? type).toString();
+  DartType get modelType => typeAnnotation?.type?.nullIfDynamic ?? initializerExpression?.staticType ?? PseudoType(typeAnnotation.toString());
 
-  String get typeModelName => type.isDynamic ? typeText : type.getDisplayString(withNullability: true);
+  String get modelTypeName => modelType.getDisplayString(withNullability: true);
 
-  // TODO: using replaceAll is a temporary hack.
-  // It is needed for now, since we cannot construct a DartType for the yet to
-  // be generated classes, ie. for _A given A. Once the new static meta
-  // programming feature is added to dart, we should be able to resolve this
-  // using a ClassTypeMacro.
-  String get typeName => typeModelName.replaceAll(session.prefix, '');
+  String get mappedTypeName => modelType.mappedName;
 
   RealmFieldInfo? get realmInfo {
     try {
@@ -82,6 +78,22 @@ extension FieldElementEx on FieldElement {
       final primaryKey = primaryKeyInfo;
       final indexed = indexedInfo;
 
+      // Check for as-of-yet unsupported type
+      if (type.isDartCoreSet || //
+          type.isDartCoreMap ||
+          type.isRealmAny ||
+          type.isExactly<Decimal128>() ||
+          type.isExactly<ObjectId>() ||
+          type.isExactly<Uuid>()) {
+        throw RealmInvalidGenerationSourceError(
+          'Field type not supported yet',
+          element: this,
+          primarySpan: typeSpan(span!.file),
+          primaryLabel: 'not yet supported',
+          todo: 'Avoid using $modelTypeName for now',
+        );
+      }
+
       // Validate primary key
       if (primaryKey != null) {
         if (type.isNullable) {
@@ -94,7 +106,7 @@ extension FieldElementEx on FieldElement {
             primaryLabel: 'is nullable',
             todo: //
                 'Consider using the @Indexed() annotation instead, '
-                "or make '$displayName' ${anOrA(typeText)} ${type.asNonNullable}.",
+                "or make '$displayName' ${anOrA(modelTypeName)} ${type.asNonNullable}.",
           );
         }
         if (indexed != null) {
@@ -106,8 +118,8 @@ extension FieldElementEx on FieldElement {
           ));
         }
         // Since the setter of a dart late final public field without initializer is public,
-        // the error of setting a primary key after construction will be a runtime error no matter 
-        // what we do. See: 
+        // the error of setting a primary key after construction will be a runtime error no matter
+        // what we do. See:
         //
         //  https://github.com/dart-lang/language/issues/1239
         //  https://github.com/dart-lang/language/issues/2068
@@ -115,7 +127,7 @@ extension FieldElementEx on FieldElement {
         // Hence we may as well lift the restriction that primary keys must be declared final.
         //
         // However, this may change in the future. Either as the dart language team change this
-        // blemish. Or perhaps we can avoid the late modifier, once static meta programming lands 
+        // blemish. Or perhaps we can avoid the late modifier, once static meta programming lands
         // in dart. Therefor we keep the code outcommented for later.
         /*
         if (!isFinal) {
@@ -146,7 +158,7 @@ extension FieldElementEx on FieldElement {
           'Realm only support indexes on String, int, and bool fields',
           element: this,
           primarySpan: typeSpan(file),
-          primaryLabel: "$typeText is not an indexable type",
+          primaryLabel: "$modelTypeName is not an indexable type",
           todo: //
               "Change the type of '$displayName', "
               "or remove the $annotation annotation",
@@ -162,10 +174,10 @@ extension FieldElementEx on FieldElement {
         String todo;
         if (notARealmTypeSpan != null) {
           todo = //
-              "Add a @RealmModel annotation on '$typeName', "
+              "Add a @RealmModel annotation on '$mappedTypeName', "
               "or an @Ignored annotation on '$displayName'.";
-        } else if (type.isDynamic && typeName != 'dynamic' && !typeName.startsWith(session.prefix)) {
-          todo = "Did you intend to use _$typeName as type for '$displayName'?";
+        } else if (type.isDynamic && mappedTypeName != 'dynamic' && !mappedTypeName.startsWith(session.prefix)) {
+          todo = "Did you intend to use _$mappedTypeName as type for '$displayName'?";
         } else {
           todo = "Remove the invalid field or add an @Ignored annotation on '$displayName'.";
         }
@@ -174,7 +186,7 @@ extension FieldElementEx on FieldElement {
           'Not a realm type',
           element: this,
           primarySpan: typeSpan(file),
-          primaryLabel: '$typeText is not a realm model type',
+          primaryLabel: '$modelTypeName is not a realm model type',
           secondarySpans: {
             modelSpan: "in realm model '${enclosingElement.displayName}'",
             // may go both above and below, or stem from another file
@@ -211,7 +223,7 @@ extension FieldElementEx on FieldElement {
               'Realm object references must be nullable',
               primarySpan: typeSpan(file),
               primaryLabel: 'is not nullable',
-              todo: 'Change type to $typeText?',
+              todo: 'Change type to $modelTypeName?',
               element: this,
             );
           }
@@ -229,7 +241,6 @@ extension FieldElementEx on FieldElement {
       rethrow;
     } catch (e, s) {
       // Fallback. Not perfect, but better than just forwarding original error.
-      print(s);
       throw RealmInvalidGenerationSourceError(
         '$e',
         todo: //