+++ CHANGELOG.md
@@ -15,6 +15,9 @@ x.x.x Release notes (yyyy-MM-dd)
 * List fields no longer required to be `final` in data model classes. ([#253](https://github.com/realm/realm-dart/pull/253))
 * Support custom FIFO special files. ([#284](https://github.com/realm/realm-dart/pull/284))
 
+### Internal
+ * Use `RealmList` instead of `List` in generated Realm object classes. It enables access to a Realm list specific APIs. ([#270](https://github.com/realm/realm-dart/pull/270))
+
 ### Compatibility
 * Dart ^2.15 on Windows, MacOS and Linux